+++ 4주차/login-signup/package.json
@@ -13,6 +13,7 @@
     "axios": "^1.6.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.19.0",
     "styled-components": "^6.1.1"
   },
   "devDependencies": {
